refactor: split view lifecycle tests
