+++ tests/helpers/obsidian-dom.js
@@ -70,6 +70,9 @@ function applyExtensions(el) {
       if (opts.cls) child.className = opts.cls;
       if (opts.text !== undefined) child.textContent = opts.text;
       if (opts.value !== undefined && 'value' in child) child.value = opts.value;
+      if (opts.type !== undefined) child.setAttribute('type', String(opts.type));
+      if (opts.placeholder !== undefined) child.setAttribute('placeholder', String(opts.placeholder));
+      if (opts.title !== undefined) child.setAttribute('title', String(opts.title));
       if (opts.href && 'href' in child) child.href = opts.href;
       if (opts.attr && typeof opts.attr === 'object') {
         Object.entries(opts.attr).forEach(([key, value]) => {
